+++ BitmovinAnalyticsCollector/Classes/Collector/analytics/BitmovinAnalyticsInternal.swift
@@ -42,7 +42,7 @@ open class BitmovinAnalyticsInternal: NSObject {
     
     deinit {
         if(self.stateMachine.state == .playing){
-            let enterTimestamp = stateMachine.enterTimestamp ?? 0
+            let enterTimestamp = stateMachine.stateEnterTimestamp
             let duration = Date().timeIntervalSince1970Millis - enterTimestamp
             if (duration < 90000) {
                 self.stateMachine.videoTimeEnd = self.adapter?.currentTime