+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - strict protocol `Collector` as public API for player collectors
+- during `paused` state we do not track `qualityChange`, `seeking` and `buffering`
 
 ### Added 
 
@@ -22,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - deprecated AVPlayerCollector
 - strict protocol `Collector` as public API for player collectors
+- during `paused` state we do not track `qualityChange`, `seeking` and `buffering`
 
 ## 2.7.0